label-reader/modules/label_extraction/requirements.txt
# Runtime
azure-iot-device==2.10.0
python-Levenshtein==0.12.2
azure-cognitiveservices-vision-computervision==0.9.0
azure-ai-formrecognizer==3.2.0
requests==2.27.1
pyclipper==1.3.0.post2
opencensus-ext-azure==1.1.1
opencensus-ext-logging==0.1.1
opencensus-ext-requests==0.7.6
rx==3.2.0
websocket-client==1.3.1
rel==0.4.7
msgpack==1.0.3
symspellpy==6.7.6
expiringdict==1.2.1
azure-storage-blob==12.11.0
tenacity==8.0.1

# Devlopment
nose2==0.11.0